requirements.txt
# AI Ready RAG - Full Requirements
# NOTE: chromadb requires onnxruntime which may not be available on all platforms
# For API-only testing, use: pip install -r requirements-api.txt

# ============== API Backend ==============
# (Always needed)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
sqlalchemy>=2.0.0
bcrypt>=4.0.0
pyjwt>=2.8.0
httpx>=0.27.0
python-dotenv

# ============== RAG Components ==============
# (Only needed for document processing and chat - skip for auth testing)

# Document Processing
docling>=2.0.0
docling-core

# LLM & Embeddings
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0

# Vector Store - requires onnxruntime
# chromadb>=0.5.0  # Uncomment on DGX Spark / Linux with onnxruntime support

# Web UI
gradio>=5.0.0

# Utilities
aiohttp
numpy

# ============== Testing ==============
pytest>=8.0.0
pytest-cov>=4.0.0
